requirements.txt
flask
requests
pandas
scikit-learn
matplotlib
python-dotenv
flask-sqlalchemy
